Arreglo error al rotar pantalla

+++ app/src/main/java/ar/edu/unc/famaf/redditreader/backend/GetTopPostsTask.java
@@ -4,6 +4,7 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.AsyncTask;
+import android.widget.Toast;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -12,6 +13,7 @@ import java.util.List;
 
 import javax.net.ssl.HttpsURLConnection;
 
+import ar.edu.unc.famaf.redditreader.R;
 import ar.edu.unc.famaf.redditreader.model.PostModel;
 import ar.edu.unc.famaf.redditreader.ui.PostsIteratorListener;
 
@@ -42,19 +44,23 @@ class    GetTopPostsTask extends AsyncTask<URL, Integer, List<PostModel>>{
     @Override
     protected void onPostExecute(List<PostModel> postModels) {
         super.onPostExecute(postModels);
-        RedditDBHelper redditDB = new RedditDBHelper(context, RedditDBHelper.POST_TABLE_VERSION);
-        SQLiteDatabase db = redditDB.getWritableDatabase();
-        db.delete(RedditDBHelper.POST_TABLE, null, null);
-        ContentValues values = new ContentValues();
-        for (PostModel postModel : postModels) {
-            values.put(RedditDBHelper.POST_TABLE_TITLE, postModel.getPostTitle());
-            values.put(RedditDBHelper.POST_TABLE_SUBREDDIT, postModel.getPostSubReddit());
-            values.put(RedditDBHelper.POST_TABLE_DATE, postModel.getPostDate());
-            values.put(RedditDBHelper.POST_TABLE_COMMENTS_COUNT, postModel.getPostCommentCount());
-            values.put(RedditDBHelper.POST_TABLE_THUMBNAIL_URL, postModel.getPostImageURL());
-            values.put(RedditDBHelper.POST_TABLE_REDDIT_ID, postModel.getPostID());
-            db.insert(RedditDBHelper.POST_TABLE, null, values);
+        if (postModels != null) {
+            RedditDBHelper redditDB = new RedditDBHelper(context, RedditDBHelper.POST_TABLE_VERSION);
+            SQLiteDatabase db = redditDB.getWritableDatabase();
+            db.delete(RedditDBHelper.POST_TABLE, null, null);
+            ContentValues values = new ContentValues();
+            for (PostModel postModel : postModels) {
+                values.put(RedditDBHelper.POST_TABLE_TITLE, postModel.getPostTitle());
+                values.put(RedditDBHelper.POST_TABLE_SUBREDDIT, postModel.getPostSubReddit());
+                values.put(RedditDBHelper.POST_TABLE_DATE, postModel.getPostDate());
+                values.put(RedditDBHelper.POST_TABLE_COMMENTS_COUNT, postModel.getPostCommentCount());
+                values.put(RedditDBHelper.POST_TABLE_THUMBNAIL_URL, postModel.getPostImageURL());
+                values.put(RedditDBHelper.POST_TABLE_REDDIT_ID, postModel.getPostID());
+                db.insert(RedditDBHelper.POST_TABLE, null, values);
+            }
+            Backend.getInstance().getNextPosts(context, postsIteratorListener, 0);
+        }else {
+            Toast.makeText(context, R.string.error_message, Toast.LENGTH_LONG).show();
         }
-        Backend.getInstance().getNextPosts(context,postsIteratorListener, 0, 0);
     }
 }
